refactor: :recycle: files cleanup
