--- src/components/CreateApplicationModal/index.css
@@ -1,6 +1,0 @@
-.newAppForm {
-  padding-top: 1rem;
-}
-.formSubmitBtn {
-  float: right;
-}
