+++ src/components/application_card/style.css
