+++ src/components/testcase_blocks/add_block/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { PlusOutlined } from "@ant-design/icons";
 import { Button, Dropdown, MenuProps } from "antd";
 import "../style.css";
-import { TEST_CASE_ITEMS } from "../index";
+import { TEST_CASE_BLOCKS } from "../index";
 import { MenuClickEventHandler } from "rc-menu/lib/interface";
 
 interface AddBlockProps {
@@ -14,19 +14,19 @@ export function AddBlock(props: AddBlockProps) {
   const items: MenuProps["items"] = [
     {
       label: "Action",
-      key: TEST_CASE_ITEMS.ACTION
+      key: TEST_CASE_BLOCKS.ACTION
     },
     {
       label: "Assert",
-      key: TEST_CASE_ITEMS.ASSERT
+      key: TEST_CASE_BLOCKS.ASSERT
     }
     // {
     //   label: "If Block",
-    //   key: TEST_CASE_ITEMS.IF
+    //   key: TEST_CASE_BLOCKS.IF
     // },
     // {
     //   label: "Iteration",
-    //   key: TEST_CASE_ITEMS.FOR_LOOP
+    //   key: TEST_CASE_BLOCKS.FOR_LOOP
     // }
   ];
