--- src/pages/action_group/style.css
@@ -1,3 +1,0 @@
-.targetContainer {
-  display: flex;
-}
